Refactor: Implement event card carousel
- Convert event cards to a horizontal carousel.
- Truncate event descriptions to two lines.
- Adjust image display to a rectangular aspect ratio.
- Modify "Learn More" to open the external link on card click.
- Abbreviate month names in date display.

+++ src/lib/utils.ts
@@ -8,7 +8,7 @@ export function cn(...inputs: ClassValue[]) {
 
 export function formatDate(dateString: string): string {
   return new Date(dateString).toLocaleDateString('en-US', {
-    month: 'long',
+    month: 'short', // Changed from 'long' to 'short' for abbreviated month names
     day: 'numeric',
     year: 'numeric'
   });
